+++ src/osm_flex/download.py
@@ -14,6 +14,7 @@ downloading functions
 import logging
 from pathlib import Path
 import urllib.request
+from urllib.parse import urljoin
 from osm_flex.config import DICT_GEOFABRIK, GEOFABRIK_URL, PLANET_URL, OSM_DATA_DIR
 
 LOGGER = logging.getLogger(__name__)
@@ -36,27 +37,60 @@ def _create_gf_download_url(iso3, file_format):
 
     Returns
     -------
-    str : Geofabrik download-string for the requested country.
+    url: str
+        Geofabrik download URL for the requested country.
 
     See also
     --------
     DICT_GEOFABRIK for exceptions / special regions.
     """
+    # Retrieve Geofabrik definitions
     try:
-        if file_format == 'shp':
-            return f'{GEOFABRIK_URL}{DICT_GEOFABRIK[iso3][0]}/{DICT_GEOFABRIK[iso3][1]}-latest-free.shp.zip'
-        if file_format == 'pbf':
-            return f'{GEOFABRIK_URL}{DICT_GEOFABRIK[iso3][0]}/{DICT_GEOFABRIK[iso3][1]}-latest.osm.pbf'
-    except KeyError:
-        if iso3=='RUS':
-            raise KeyError("""Russia comes in two files. Please specify either
-                         'RUS-A for the Asian or RUS-E for the European part.""")
-        raise KeyError("""The provided iso3 seems not to be available on
-                           Geofabrik.de. You can clip it from the planet
-                           file or an adequate regional file, instead. See
-                           the methods in the clip module for this.""")
+        continent, country = DICT_GEOFABRIK[iso3]
 
-    return LOGGER.error('invalid file format. Please choose one of [shp, pbf]')
+    except KeyError as err:
+        if iso3 == "RUS":
+            raise KeyError(
+                "Russia comes in two files. Please specify either RUS-A for the Asian "
+                "or RUS-E for the European part."
+            ) from err
+        raise KeyError(
+            "The provided iso3 seems not to be available on Geofabrik.de. You can clip "
+            "it from the planet file or an adequate regional file, instead. See the "
+            "methods in the clip module for this."
+        ) from err
+
+    # Set file extension
+    if file_format == "shp":
+        ext = "-free.shp.zip"
+    elif file_format == "pbf":
+        ext = ".osm.pbf"
+    else:
+        raise NotImplementedError(
+            f"Invalid file format '{file_format}'. Please choose one of [shp, pbf]"
+        )
+
+    # Join to URL
+    return urljoin(GEOFABRIK_URL, f"{continent}/{country}-latest{ext}")
+
+def _download_file(download_url: str, filepath: Path, overwrite: bool = True):
+    """Download a file located at an URL to a local file path
+
+    Parameters
+    ----------
+    download_url : str
+        URL of the file to download
+    filepath : str or Path
+        Local file path to store the file
+    overwrite : bool, optional
+        Overwrite existing files. If ``False``, the download will be skipped for
+        existing files. Defaults to ``True``.
+    """
+    if not Path(filepath).is_file() or overwrite:
+        LOGGER.info(f"Download file: {filepath}")
+        urllib.request.urlretrieve(download_url, filepath)
+    else:
+        LOGGER.info(f"Skip existing file: {filepath}")
 
 # TODO: decide whether to issue warnings for multi-country files
 def get_country_geofabrik(iso3, file_format='pbf', save_path=OSM_DATA_DIR,
@@ -82,8 +116,8 @@ def get_country_geofabrik(iso3, file_format='pbf', save_path=OSM_DATA_DIR,
 
     Returns
     -------
-    None
-        File is downloaded and stored under save_path + the Geofabrik filename
+    filepath : Path
+        The path to the downloaded file (``save_path`` + the Geofabrik filename)
 
     See also
     --------
@@ -91,12 +125,10 @@ def get_country_geofabrik(iso3, file_format='pbf', save_path=OSM_DATA_DIR,
     """
 
     download_url = _create_gf_download_url(iso3, file_format)
-    local_filepath = Path(save_path , download_url.split('/')[-1])
-    if not Path(local_filepath).is_file() or overwrite:
-        LOGGER.info(f'Downloading file as {local_filepath}')
-        urllib.request.urlretrieve(download_url, local_filepath)
-    else:
-        LOGGER.info(f'file already exists as {local_filepath}')
+    filepath = Path(save_path, Path(download_url).name)
+    _download_file(download_url, filepath, overwrite)
+
+    return filepath
 
 # TODO: allow for several spelling options like "Central America", "Australia", ...
 def get_region_geofabrik(region, save_path=OSM_DATA_DIR, overwrite=False):
@@ -111,15 +143,18 @@ def get_region_geofabrik(region, save_path=OSM_DATA_DIR, overwrite=False):
         Central-America, Europe, North-America, South-America
     save_path : str or pathlib.Path
         Folder in which to save the file
+
+    Returns
+    -------
+    filepath : Path
+        The path to the downloaded file
     """
 
     download_url =  f'{GEOFABRIK_URL}{region.lower()}-latest.osm.pbf'
-    local_filepath = Path(save_path , download_url.split('/')[-1])
-    if not Path(local_filepath).is_file() or overwrite:
-        LOGGER.info(f'Downloading file as {local_filepath}')
-        urllib.request.urlretrieve(download_url, local_filepath)
-    else:
-        LOGGER.info(f'file already exists as {local_filepath}')
+    filepath = Path(save_path, Path(download_url).name)
+    _download_file(download_url, filepath, overwrite)
+
+    return filepath
 
 
 def get_planet_file(save_path=Path(OSM_DATA_DIR,'planet-latest.osm.pbf'),
@@ -130,10 +165,13 @@ def get_planet_file(save_path=Path(OSM_DATA_DIR,'planet-latest.osm.pbf'),
     Parameters
     ----------
     save_path : str or pathlib.Path
-    """
+        The path to store the file.
 
-    if not Path(save_path).is_file() or overwrite:
-        LOGGER.info(f'Downloading file as {save_path}')
-        urllib.request.urlretrieve(PLANET_URL, save_path)
-    else:
-        LOGGER.info(f'file already exists as {save_path}')
+    Returns
+    -------
+    save_path : Path
+        The path to the downloaded file. Returned for consistency with other download
+        functions.
+    """
+    _download_file(PLANET_URL, save_path, overwrite)
+    return Path(save_path)